Notification Personalisation Refactor
- Remove each individual Personalisation template
- Create data from generic Personalisation so we do not need to pick and choose fields for the template. Gov Notify will handle this
