+++ src/main/java/uk/gov/hmcts/sscs/factory/PersonalisationFactory.java
@@ -1,21 +1,14 @@
 package uk.gov.hmcts.sscs.factory;
 
-import static org.slf4j.LoggerFactory.getLogger;
-import static uk.gov.hmcts.sscs.domain.notify.NotificationType.ADJOURNED;
-
 import java.util.function.Function;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.sscs.config.NotificationConfig;
-import uk.gov.hmcts.sscs.controller.NotificationController;
 import uk.gov.hmcts.sscs.domain.notify.NotificationType;
-import uk.gov.hmcts.sscs.exception.NotificationClientRuntimeException;
-import uk.gov.hmcts.sscs.placeholders.*;
+import uk.gov.hmcts.sscs.personalisation.Personalisation;
 
 @Component
 public class PersonalisationFactory implements Function<NotificationType, Personalisation> {
-
-    private static final org.slf4j.Logger LOG = getLogger(NotificationController.class);
 
     private final NotificationConfig config;
 
@@ -27,22 +20,7 @@ public class PersonalisationFactory implements Function<NotificationType, Person
     @Override
     public Personalisation apply(NotificationType notificationType) {
         if (notificationType != null) {
-            switch (notificationType) {
-                case ADJOURNED:
-                    return new DefaultPersonalisation(config, ADJOURNED);
-                case APPEAL_RECEIVED:
-                    return new AppealReceivedPersonalisation(config);
-                case DWP_RESPONSE_RECEIVED:
-                    return new ResponseReceivedPersonalisation(config);
-                case EVIDENCE_RECEIVED:
-                    return new EvidenceReceivedPersonalisation(config);
-                case POSTPONEMENT:
-                    return new PostponementPersonalisation(config);
-                default:
-                    String error = "Unknown Notification type received: " + notificationType;
-                    LOG.error(error);
-                    throw new NotificationClientRuntimeException(new Exception(error));
-            }
+            return new Personalisation(config);
         }
         return null;
     }
