+++ src/main/java/uk/gov/hmcts/sscs/personalisation/Personalisation.java
@@ -1,16 +1,19 @@
-package uk.gov.hmcts.sscs.placeholders;
+package uk.gov.hmcts.sscs.personalisation;
 
+import static java.time.temporal.ChronoUnit.DAYS;
 import static uk.gov.hmcts.sscs.config.AppConstants.*;
 
+import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.Map;
 import uk.gov.hmcts.sscs.config.NotificationConfig;
 import uk.gov.hmcts.sscs.domain.CcdResponse;
+import uk.gov.hmcts.sscs.domain.notify.NotificationType;
 import uk.gov.hmcts.sscs.domain.notify.Template;
 
-public abstract class Personalisation {
+public class Personalisation {
 
     protected NotificationConfig config;
 
@@ -30,8 +33,21 @@ public abstract class Personalisation {
         //TODO: Replace hardcoded mactoken with an actual mac token
         personalisation.put(MANAGE_EMAILS_LINK_LITERAL, config.getManageEmailsLink().replace(MAC_LITERAL, "Mactoken"));
         personalisation.put(TRACK_APPEAL_LINK_LITERAL, config.getTrackAppealLink() != null ? config.getTrackAppealLink().replace(APPEAL_ID_LITERAL, ccdResponse.getAppellantSubscription().getAppealNumber()) : null);
+        personalisation.put(FIRST_TIER_AGENCY_ACRONYM, DWP_ACRONYM);
+        personalisation.put(FIRST_TIER_AGENCY_FULL_NAME, DWP_FUL_NAME);
+        // TODO: Set this to the actual event date once event story has been implemented
+        ZonedDateTime dwpResponseDate = ZonedDateTime.of(1900, 1, 1, 0, 0, 0, 0, ZoneId.of("GMT"));
+        String dwpResponseDateString = formatDate(dwpResponseDate.plus(MAX_DWP_RESPONSE_DAYS, DAYS));
+        personalisation.put(APPEAL_RESPOND_DATE, dwpResponseDateString);
+        // TODO: Set this to the actual event date once event story has been implemented
+        ZonedDateTime evidenceReceivedDate = ZonedDateTime.of(1900, 1, 1, 0, 0, 0, 0, ZoneId.of("GMT"));
+        String evidenceReceivedDateString = formatDate(evidenceReceivedDate);
+        personalisation.put(EVIDENCE_RECEIVED_DATE_LITERAL, evidenceReceivedDateString);
+        // TODO: Set this to the actual event date once event story has been implemented
+        ZonedDateTime z = ZonedDateTime.of(1900, 1, 1, 0, 0, 0, 0, ZoneId.of("GMT"));
+        personalisation.put(HEARING_CONTACT_DATE, formatDate(z.plusWeeks(6)));
+        personalisation.put(SUBMIT_EVIDENCE_LINK_LITERAL, config.getEvidenceSubmissionInfoLink().replace(APPEAL_ID, ccdResponse.getAppellantSubscription().getAppealNumber()));
 
-        personalisation = customise(ccdResponse, personalisation);
         return personalisation;
     }
 
@@ -39,7 +55,7 @@ public abstract class Personalisation {
         return date.format(DateTimeFormatter.ofPattern(RESPONSE_DATE_FORMAT));
     }
 
-    protected abstract Map<String, String> customise(CcdResponse event, Map<String, String> personalisation);
-
-    public abstract Template getTemplate();
+    public Template getTemplate(NotificationType type) {
+        return config.getTemplate(type.getId());
+    }
 }
